pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Owen Hoskin <hoskinowen7gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
waitress = "^2.1.2"
Flask = "^2.1.2"
requests = "^2.28.1"
aiofiles = "^0.8.0"
aiocsv = "^1.2.2"
"discord.py" = {extras = ["speed"], version = "^2.0.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"